requirements.txt
hydra-core
omegaconf
openai
transformers
torch
spacy
faiss-cpu
pyserini
numpy
tqdm